Change value data type

+++ app/data/models/completed-payment-request.js
@@ -16,7 +16,7 @@ module.exports = (sequelize, DataTypes) => {
     currency: DataTypes.STRING,
     schedule: DataTypes.STRING,
     dueDate: DataTypes.STRING,
-    value: DataTypes.DECIMAL,
+    value: DataTypes.INTEGER,
     acknowledged: DataTypes.DATE,
     settled: DataTypes.DATE
   },
